requirements.txt
flask>=2.0
flask-limiter>=3.0
flasgger>=0.9.5
marshmallow>=3.0
psutil>=5.8
requests>=2.0
pytest>=7.0
pystray>=0.19
pillow>=9.0